requirements.txt
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
transformers>=4.30.0  # For Wav2Vec2

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
h5py>=3.8.0

# Video Processing
# decord>=0.6.0  # Efficient video loading (not available for Python 3.13)
# pytorchvideo>=0.1.5  # Optional

# Reinforcement Learning
gym>=0.26.0
stable-baselines3>=2.0.0  # Optional, for baseline comparison

# Experiment Tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Configuration
PyYAML>=6.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# API & Deployment
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Model Optimization
onnx>=1.14.0
onnxruntime>=1.15.0
